refactor: implement repository pattern and move Note-related classes to data.model package
