+++ app/src/main/java/com/ragdoll/notesapplication/data/NoteRepository.kt
@@ -1,0 +1,15 @@
+package com.ragdoll.notesapplication.data
+
+import androidx.lifecycle.LiveData
+import com.ragdoll.notesapplication.data.model.Note
+
+interface NoteRepository {
+    fun upsertNote(note: Note)
+
+    fun getNotes(): LiveData<List<Note>>
+
+    fun deleteNote(note: Note)
+
+    fun deleteAllNotes()
+
+}
