+++ app/src/main/java/com/ragdoll/notesapplication/data/model/Note.kt
@@ -1,4 +1,4 @@
-package com.ragdoll.notesapplication
+package com.ragdoll.notesapplication.data.model
 
 import android.os.Parcelable
 import androidx.room.ColumnInfo
